requirements.txt
Django==1.11.12
django-debug-toolbar==1.9.1
django-extensions==2.0.0
django-filter==1.0.1
django-rest-auth==0.9.3
djangorestframework==3.8.2
pkg-resources==0.0.0
psycopg2==2.7.3.1
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
typing==3.6.6
